# api/openapi.json
{
  "components": {
    "schemas": {
...
      "Catalog": {
        "additionalProperties": false,
        "properties": {
          "ages": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "civilizations": {
            "items": {
              "$ref": "#/components/schemas/Civilization"
            },
            "type": "array"
          },
          "definitions": {
            "items": {
              "$ref": "#/components/schemas/Definition"
            },
            "type": "array"
          },
          "difficulties": {
            "items": {
              "$ref": "#/components/schemas/Difficulty"
            },
            "type": "array"
          },
          "log_filters": {
            "items": {
              "$ref": "#/components/schemas/LogFilter"
            },
            "type": "array"
          },
          "rules_version": {
            "type": "string"
          },
          "settlement_counts": {
            "items": {
              "type": "integer"
            },
            "type": "array"
          },
          "speeds": {
            "items": {
              "type": "number"
            },
            "type": "array"
          },
          "technologies": {
            "items": {
              "$ref": "#/components/schemas/Technology"
            },
            "type": "array"
          }
        },
        "required": [
          "rules_version",
          "definitions",
          "technologies",
          "civilizations",
          "ages",
          "speeds",
          "difficulties",
          "log_filters",
          "settlement_counts"
        ],
        "type": "object"
      },
...
      "Command": {
        "additionalProperties": false,
        "properties": {
          "entity_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array"
          },
          "id": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "position": {
            "$ref": "#/components/schemas/Vec"
          },
          "product": {
            "type": "string"
          },
          "queue": {
            "type": "boolean"
          },
          "target_id": {
            "type": "integer"
          },
          "target_player": {
            "type": "integer"
          },
          "value": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "kind"
        ],
        "type": "object"
      },
      "Config": {
        "additionalProperties": false,
        "properties": {
          "civilization": {
            "type": "string"
          },
          "difficulty": {
            "type": "string"
          },
          "mode": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "seed": {
            "type": "integer"
          },
          "settlements": {
            "type": "integer"
          }
        },
        "required": [
          "seed",
          "civilization",
          "difficulty",
          "mode"
        ],
        "type": "object"
      },
...
      "ErrorBody": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/RuleError"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "Session": {
        "additionalProperties": false,
        "properties": {
          "match_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "player_id": {
            "type": "integer"
          },
          "token": {
            "type": "string"
          }
        },
        "required": [
          "match_id",
          "token",
          "player_id",
          "name"
        ],
        "type": "object"
      },
      "Snapshot": {
        "additionalProperties": false,
        "properties": {
          "build_options": {
            "items": {
              "$ref": "#/components/schemas/Action"
            },
            "type": "array"
          },
          "difficulty": {
            "$ref": "#/components/schemas/Difficulty"
          },
          "entities": {
            "items": {
              "$ref": "#/components/schemas/EntityView"
            },
            "type": "array"
          },
          "event_cursor": {
            "type": "integer"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/Event"
            },
            "type": "array"
          },
          "map": {
            "$ref": "#/components/schemas/MapView"
          },
          "opponents": {
            "items": {
              "$ref": "#/components/schemas/OpponentView"
            },
            "type": "array"
          },
          "paused": {
            "type": "boolean"
          },
          "player": {
            "$ref": "#/components/schemas/PlayerView"
          },
          "projectiles": {
            "items": {
              "$ref": "#/components/schemas/ProjectileView"
            },
            "type": "array"
          },
          "settlements": {
            "type": "integer"
          },
          "speed": {
            "type": "number"
          },
          "status": {
            "type": "string"
          },
          "tick": {
            "type": "integer"
          },
          "time": {
            "type": "number"
          },
          "version": {
            "type": "string"
          },
          "winner": {
            "type": "integer"
          }
        },
        "required": [
          "version",
          "difficulty",
          "settlements",
          "tick",
          "time",
          "speed",
          "paused",
          "status",
          "winner",
          "player",
          "opponents",
          "map",
          "entities",
          "projectiles",
          "events",
          "event_cursor",
          "build_options"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "matchToken": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Go-authoritative RTS. Commands are intent; snapshots are authorized presentation state.",
    "title": "AI of Empires API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/catalog": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Catalog"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Structured error"
          }
        },
        "summary": "Read display metadata and ruleset catalog"
      }
    },
...
    "/matches": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Config"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Structured error"
          }
        },
        "summary": "Create a single-player match against server AI"
      }
    },
    "/matches/{id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Structured error"
          }
        },
        "security": [
          {
            "matchToken": []
          }
        ],
        "summary": "Permanently delete a game and its saved history"
      },
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Snapshot"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Structured error"
          }
        },
        "security": [
          {
            "matchToken": []
          }
        ],
        "summary": "Read your fog-filtered snapshot"
      }
    },
    "/matches/{id}/commands": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Command"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Receipt"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Structured error"
          }
        },
        "security": [
          {
            "matchToken": []
          }
        ],
        "summary": "Submit an idempotent intention; the server validates all rules"
      }
    },
    "/matches/{id}/entities/{entity}/history": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "entity",
            "required": true,
            "schema": {
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "after and before are exclusive cursors and mutually exclusive; after=0 starts at the beginning",
            "in": "query",
            "name": "after",
            "schema": {
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "description": "after and before are exclusive cursors and mutually exclusive; after=0 starts at the beginning",
            "in": "query",
            "name": "before",
            "schema": {
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "after and before are exclusive cursors and mutually exclusive; after=0 starts at the beginning",
            "in": "query",
            "name": "limit",
            "schema": {
              "default": 100,
              "maximum": 200,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Case-insensitive search of all retained authorized events; all words must match. Identity forms Villager 3, #3, or a lone 3 match the exact entity ID. Other numbers search event text.",
            "in": "query",
            "name": "q",
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          },
          {
            "description": "Combine a catalog event category with search and pagination. Cursors and availability refer to matching events.",
            "in": "query",
            "name": "category",
            "schema": {
              "enum": [
                "",
                "orders",
                "activity",
                "economy",
                "combat",
                "production",
                "world"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventPage"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Structured error"
          }
        },
        "security": [
          {
            "matchToken": []
          }
        ],
        "summary": "Read immutable events visible when they occurred; ascending IDs, newest page by default"
      }
    },
    "/matches/{id}/events": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "SSE event `snapshot`, id = simulation tick; JSON data conforms to Snapshot. Bearer auth through fetch streaming."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Structured error"
          }
        },
        "security": [
          {
            "matchToken": []
          }
        ],
        "summary": "Stream complete fog-filtered snapshots; reconnect replaces state"
      }
    },
    "/matches/{id}/leave": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedGame"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Structured error"
          }
        },
        "security": [
          {
            "matchToken": []
          }
        ],
        "summary": "Checkpoint and unload a session until it is resumed"
      }
    },
    "/matches/{id}/log": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "after and before are exclusive cursors and mutually exclusive; after=0 starts at the beginning",
            "in": "query",
            "name": "after",
            "schema": {
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "description": "after and before are exclusive cursors and mutually exclusive; after=0 starts at the beginning",
            "in": "query",
            "name": "before",
            "schema": {
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "after and before are exclusive cursors and mutually exclusive; after=0 starts at the beginning",
            "in": "query",
            "name": "limit",
            "schema": {
              "default": 100,
              "maximum": 200,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Case-insensitive search of all retained authorized events; all words must match. Identity forms Villager 3, #3, or a lone 3 match the exact entity ID. Other numbers search event text.",
            "in": "query",
            "name": "q",
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          },
          {
            "description": "Combine a catalog event category with search and pagination. Cursors and availability refer to matching events.",
            "in": "query",
            "name": "category",
            "schema": {
              "enum": [
                "",
                "orders",
                "activity",
                "economy",
                "combat",
                "production",
                "world"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventPage"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Structured error"
          }
        },
        "security": [
          {
            "matchToken": []
          }
        ],
        "summary": "Read immutable events visible when they occurred; ascending IDs, newest page by default"
      }
    },
    "/matches/{id}/placement": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Placement"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlacementResult"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Structured error"
          }
        },
        "security": [
          {
            "matchToken": []
          }
        ],
        "summary": "Check a proposed building site on the server"
      }
    },
    "/matches/{id}/save": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedGame"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Structured error"
          }
        },
        "security": [
          {
            "matchToken": []
          }
        ],
        "summary": "Atomically checkpoint the simulation, receipts and history"
      }
    },
    "/matches/{id}/session": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedGame"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Structured error"
          }
        },
        "security": [
          {
            "matchToken": []
          }
        ],
        "summary": "Read game name and last successful autosave"
      }
    },
    "/sessions": {
      "get": {
        "parameters": [
          {
            "description": "Case-insensitive substring of a game name or session ID",
            "in": "query",
            "name": "q",
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedGames"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Structured error"
          }
        },
        "summary": "List up to 100 local saved games, newest save first"
      }
    },
    "/sessions/resume": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Structured error"
          }
        },
        "summary": "Resume a local game by exact name or ID; rotates its bearer token"
      }
    }
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ]
}